Avoid deprecated from and to parameters for range queries.
The “from” and “to” parameters for range queries were deprecated in
Elasticsearch 0.90.4, and more than ten years later, they were finally
removed in Elasticsearch 9.x.

Therefore, the existing code did not work any longer when using modern
versions of Elasticsearch. As a side effect, using the “gte” and “lte”
parameters simplifiese the code, because we can now simply tell
Elasticsearch that the values represent seconds since epoch instead of
having to convert to milliseconds in the code.